管理者画面の大幅な改修とそのサンプルとしてUser2Hogoshaマスタメンテ一式

+++ app/Http/Controllers/HogoshaController.php
@@ -21,6 +21,8 @@ use App\Consts\AuthConst;
 
 use App\Notifications\User2HogoshaRegisteredNotification;
 
+use App\Http\Requests\User2HogoshaRequest;
+
 class HogoshaController extends Controller
 {
   
@@ -163,12 +165,9 @@ class HogoshaController extends Controller
 
         return redirect()->route('hogosha-add',$args)->with('alertComp',MessageConst::DELETE_COMPLETED);
     }
-    //user2保護者登録画面へ /user2hogosha/add/
-    public function u2hadd(Request $request, Response $response){
-        
-        $items = User2Hogosha::getu2hData();
+    //user2保護者登録第一画面へ /user2hogosha/list/
+    public function u2hlist(Request $request, Response $response){
 
-        $itemsHogosha = Hogosha::all();
         //リダイレクト時には、セッションにalertが入ってくる可能性があるので拾う
         $alertComp='';
         if($request->session()->has('alertComp')){
@@ -180,23 +179,53 @@ class HogoshaController extends Controller
         }
         
         $args=[
-            'items'=>$items,
-            'itemsHogosha' =>$itemsHogosha,
             'alertComp'=>$alertComp,
             'alertErr'=>$alertErr,
         ];
 
-        return view('user2hogosha.add',$args);
+        return view('user2hogosha.list',$args);
+
+    }
+    //user2保護者登録第二画面へ /user2hogosha/edit/ GET
+    public function u2hedit(Request $request, Response $response){
+        $mode = 'create';
+
+        //u2h_idがあれば、編集モード、なければ新規モード
+        if($request->has('u2h_id')){
+            //当マスタは編集モード非対応のため
+            $mode='delete';
+        }
+        
+        //リダイレクト時には、セッションにalertが入ってくる可能性があるので拾う
+        $alertComp='';
+        if($request->session()->has('alertComp')){
+            $alertComp = $request->session()->get('alertComp');
+        }
+        $alertErr='';
+        if($request->session()->has('alertErr')){
+            $alertErr = $request->session()->get('alertErr');
+        }
+        
+        $args=[
+            'mode'=>$mode,
+            'createAction' =>"/user2hogosha/create",
+            'updateAction' =>"",
+            'deleteAction' =>"/user2hogosha/delete",
+            'backURL'=>"/user2hogosha/list/",
+            'alertComp'=>$alertComp,
+            'alertErr'=>$alertErr,
+        ];
+
+        return view('user2hogosha.edit',$args);
 
     }
     //user2保護者登録画面のPOST
-    public function u2hcreate(Request $request){
-        $this->validate($request, User2Hogosha::$rules);
+    public function u2hcreate(User2HogoshaRequest $request){
+        // $this->validate($request, User2Hogosha::$rules);
         $u2h = new User2Hogosha;
         $form = $request->all();
         unset($form['_token']);
         $u2h->fill($form);
-
 
         $u2h->save();
 
@@ -210,19 +239,21 @@ class HogoshaController extends Controller
         $args=[
         ];
 
-        return redirect()->route('u2h-add',$args)->with('alertComp',MessageConst::ADD_COMPLETED);
+        return redirect()->route('u2h-list',$args)->with('alertComp',MessageConst::ADD_COMPLETED);
 
     }
-    //
+    //user2保護者削除のPOST
     public function u2hdelete(Request $request){
         $u2h_id = $request->u2h_id;
 
         $u2h = User2Hogosha::find($u2h_id);
-        $u2h->delete();
+        if($u2h){
+            $u2h->delete();
+        }
 
         $args=[
         ];
-        return redirect()->route('u2h-add',$args)->with('alertComp',MessageConst::DELETE_COMPLETED);
+        return redirect()->route('u2h-list',$args)->with('alertComp',MessageConst::DELETE_COMPLETED);
     }
 
 }
